Add RHDisplayCategoryEventsBase
